refactor(forms): fix various styling issues
- Last-Login-IP: Inserting a minimum width for the mat card item
 
- Navbar: Correction of a shopping cart display error on small displays

Sidenav: Since on small displays the toolbar with the reintroduced shopping cart was too full and the search has shifted over a part of the content, the account menu on small displays is now in the Sidenav. This should make the search work as usual again. The account itself is displayed in the Sidenav as a nested menu.

-Two-Factor-Auth: Inserting a minimum width for the mat card item